ltlib.config - add function: get
Implement ``get``, which takes the name of a config, an optional account
context and an optional default.

If an account is given, first checks the account space for the name.
If no account given, or if the name not found in the account space,
look for the name in the global config space.  If still not found,
return the default, if given, otherwise ``None``.

Refactor functions ``reader`` and ``readerargs`` to use ``get``.